+++ auth/exceptions.go
